+++ apps/jumpserver/asgi.py
@@ -1,21 +1,8 @@
 import os
 
-from channels.auth import AuthMiddlewareStack
-from channels.routing import ProtocolTypeRouter, URLRouter
-from django.core.asgi import get_asgi_application
-
-from .middleware import WsSignatureAuthMiddleware
-from .routing import urlpatterns
+import django
+from channels.routing import get_default_application
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "jumpserver.settings")
-application = ProtocolTypeRouter({
-    # Django's ASGI application to handle traditional HTTP requests
-    "http": get_asgi_application(),
-
-    # WebSocket chat handler
-    "websocket": WsSignatureAuthMiddleware(
-        AuthMiddlewareStack(
-            URLRouter(urlpatterns)
-        )
-    ),
-})
+django.setup()
+application = get_default_application()
