+++ apps/jumpserver/middleware.py
@@ -6,16 +6,12 @@ import re
 import time
 
 import pytz
-from channels.db import database_sync_to_async
 from django.conf import settings
 from django.core.exceptions import MiddlewareNotUsed
-from django.core.handlers.asgi import ASGIRequest
 from django.http.response import HttpResponseForbidden
 from django.shortcuts import HttpResponse
 from django.utils import timezone
 
-from authentication.backends.drf import (SignatureAuthentication,
-                                         AccessTokenAuthentication)
 from .utils import set_current_request
 
 
@@ -146,35 +142,3 @@ class EndMiddleware:
         response = self.get_response(request)
         request._e_time_end = time.time()
         return response
-
-
-@database_sync_to_async
-def get_signature_user(scope):
-    headers = dict(scope["headers"])
-    if not headers.get(b'authorization'):
-        return
-    if scope['type'] == 'websocket':
-        scope['method'] = 'GET'
-    try:
-        # 因为 ws 使用的是 scope，所以需要转换成 request 对象，用于认证校验
-        request = ASGIRequest(scope, None)
-        backends = [SignatureAuthentication(),
-                    AccessTokenAuthentication()]
-        for backend in backends:
-            user, _ = backend.authenticate(request)
-            if user:
-                return user
-    except Exception as e:
-        print(e)
-    return None
-
-
-class WsSignatureAuthMiddleware:
-    def __init__(self, app):
-        self.app = app
-
-    async def __call__(self, scope, receive, send):
-        user = await get_signature_user(scope)
-        if user:
-            scope['user'] = user
-        return await self.app(scope, receive, send)
